adjust config after load of core (plus new host headers)
